cleanup preset util
