+++ examples/react/package.json
@@ -13,7 +13,7 @@
     "next": "13.0.5",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "wagmi": "0.8.7"
+    "wagmi": "0.8.8"
   },
   "devDependencies": {
     "@types/node": "18.11.9",
