+++ package.json
@@ -36,7 +36,7 @@
     "esbuild": "0.15.16",
     "eslint": "8.28.0",
     "eslint-config-prettier": "8.5.0",
-    "lerna": "6.0.3",
+    "lerna": "6.1.0",
     "prettier": "2.8.0",
     "rollup": "3.5.0",
     "rollup-plugin-esbuild": "5.0.0",
