+++ packages/ui/src/components/w3m-wallet-image/index.ts
@@ -1,6 +1,6 @@
 import { html, LitElement } from 'lit'
 import { customElement, property } from 'lit/decorators.js'
-import { getOptimisticWalletIdPreset } from '../../utils/Presets'
+import { PresetUtil } from '../../utils/PresetUtil'
 import { WALLET_PLACEHOLDER } from '../../utils/Svgs'
 import { global } from '../../utils/Theme'
 import { getWalletIcon } from '../../utils/UiHelpers'
@@ -17,7 +17,7 @@ export class W3mWalletImage extends LitElement {
   // -- render ------------------------------------------------------- //
   protected render() {
     const walletId = this.walletId ?? 'injected'
-    const optimisticId = getOptimisticWalletIdPreset(walletId)
+    const optimisticId = PresetUtil.optimisticWalletId(walletId)
     const src = this.src ? this.src : getWalletIcon(optimisticId)
 
     return html`
